チャット一覧画面における条件分岐の追加

+++ resources/views/messages/index.blade.php
@@ -6,9 +6,15 @@
       <ul class="list-group">
         @foreach($groups as $group)
           <li class="list-group-item">
-            {{ $group->textbook->sellBook->name }}（{{ $group->textbook->name }}）
-            <a href="/message/{{ $group->id }}/detail" class="btn btn-outline-success message-btn">チャット</a>
-            <a href="/message/{{ $group->id }}/delete" class="btn btn-outline-danger">終了</a>
+            @if($user_id === $group->textbook->sellBook->id)
+              {{ $group->textbook->buyBook->name }}（{{ $group->textbook->name }}）
+              <a href="/message/{{ $group->id }}/detail" class="btn btn-outline-success message-btn">チャット</a>
+              <a href="/message/{{ $group->id }}/delete" class="btn btn-outline-danger">終了</a>
+            @else
+              {{ $group->textbook->sellBook->name }}（{{ $group->textbook->name }}）
+              <a href="/message/{{ $group->id }}/detail" class="btn btn-outline-success message-btn">チャット</a>
+              <a href="/message/{{ $group->id }}/delete" class="btn btn-outline-danger">終了</a>
+            @endif
           </li>
         @endforeach
       </ul>
